remove psql project and move queryverify to unittest
