+++ qpmodel.sln
@@ -10,9 +10,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "test", "test\test.csproj", 
 		{E06A2982-3807-40DC-B025-878745E5C98A} = {E06A2982-3807-40DC-B025-878745E5C98A}
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "psql", "psql\psql.csproj", "{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "server", "server\server.csproj", "{ECE3647D-9C3F-47F7-A81F-4F4E3C6D2694}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "server", "server\server.csproj", "{ECE3647D-9C3F-47F7-A81F-4F4E3C6D2694}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -38,14 +36,6 @@ Global
 		{0E5D07A1-1915-409F-BFD5-3C7086A05926}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0E5D07A1-1915-409F-BFD5-3C7086A05926}.Release|x64.ActiveCfg = Release|Any CPU
 		{0E5D07A1-1915-409F-BFD5-3C7086A05926}.Release|x64.Build.0 = Release|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Debug|x64.Build.0 = Debug|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Release|x64.ActiveCfg = Release|Any CPU
-		{50BB43FB-34F4-4892-BFAB-2D80EF2EA6DD}.Release|x64.Build.0 = Release|Any CPU
 		{ECE3647D-9C3F-47F7-A81F-4F4E3C6D2694}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ECE3647D-9C3F-47F7-A81F-4F4E3C6D2694}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ECE3647D-9C3F-47F7-A81F-4F4E3C6D2694}.Debug|x64.ActiveCfg = Debug|Any CPU
