+++ test/UnitTest.cs
@@ -43,8 +43,6 @@ using qpmodel.dml;
 using qpmodel.tools;
 using qpmodel.stat;
 
-using psql;
-
 namespace qpmodel.unittest
 {
     // Test Utils
@@ -146,6 +144,68 @@ namespace qpmodel.unittest
                 }
             }
             return true;
+        }
+    }
+
+    public class QueryVerify
+    {
+        static bool resultVerify(string resultFn, string expectFn)
+        {
+            // read  strings from result file and expected file
+            string expectText = File.ReadAllText(expectFn).Replace("\r\n", "\n");
+            string resultText = File.ReadAllText(resultFn).Replace("\r\n", "\n");
+
+            // compare the test result with the expected result
+            if (string.Compare(resultText, expectText) != 0)
+            {
+                return false;
+            }
+
+            return true;
+        }
+
+        public string SQLQueryVerify(string sql_dir_fn, string write_dir_fn, string expect_dir_fn, string[] badQueries, bool explainOnly)
+        {
+            string result = null;
+            QueryOption option = new QueryOption();
+            option.optimize_.TurnOnAllOptimizations();
+            option.optimize_.remove_from_ = false;
+
+            option.explain_.show_output_ = true;
+            option.explain_.show_estCost_ = option.optimize_.use_memo_;
+            option.explain_.mode_ = explainOnly ? ExplainMode.explain : ExplainMode.full;
+
+            // get a list of sql query fine names from the sql directory
+            string[] sqlFiles = Directory.GetFiles(sql_dir_fn, "*.sql");
+
+            // execute the query in each file and and verify the result
+            foreach (string sqlFn in sqlFiles)
+            {
+                string dbg_name = Path.GetFileNameWithoutExtension(sqlFn);
+
+                if (badQueries.Contains(dbg_name) == true)
+                    continue;
+
+                // execute query
+                var sql = File.ReadAllText(sqlFn);
+                var test_result = SQLStatement.ExecSQLList(sql, option);
+
+                // construct file name for result file and write result
+                string f_name = Path.GetFileNameWithoutExtension(sqlFn);
+                string write_fn = $@"{write_dir_fn}/{f_name}.txt";
+
+                File.WriteAllText(write_fn, test_result);
+
+                // construct file name of expected result
+                string expect_fn = $@"{expect_dir_fn}/{f_name}.txt";
+
+                // verify query result against the expected result
+                if (!resultVerify(write_fn, expect_fn))
+                {
+                    result += write_fn + ";";
+                }
+            }
+            return result;
         }
     }
